+++ lib/src/mock.rs
@@ -59,3 +59,40 @@ pub(crate) const KEY_0: &str = "data";
 pub(crate) const KEY_1: &str = "items";
 pub(crate) const KEY_2: &str = "profile";
 pub(crate) const KEY_3: &str = "name";
+
+use std::collections::HashMap;
+
+use crate::{
+  http::{Manifest, Request, Response, ResponseBody},
+  json::JsonKey,
+};
+
+#[cfg(test)]
+pub(crate) fn mock_manifest() -> Manifest {
+  let request = Request {
+    method:  "GET".to_string(),
+    url:     "spotify.com".to_string(),
+    version: "HTTP/1.1".to_string(),
+    headers: HashMap::new(),
+  };
+  let mut headers = HashMap::new();
+  headers.insert("content-type".to_string(), "application/json; charset=utf-8".to_string());
+  headers.insert("content-encoding".to_string(), "gzip".to_string());
+  let body = ResponseBody {
+    json: vec![
+      JsonKey::String("data".to_string()),
+      JsonKey::String("items".to_string()),
+      JsonKey::Num(0),
+      JsonKey::String("profile".to_string()),
+      JsonKey::String("name".to_string()),
+    ],
+  };
+  let response = Response {
+    status: "200".to_string(),
+    version: "HTTP/1.1".to_string(),
+    message: "OK".to_string(),
+    headers,
+    body,
+  };
+  Manifest { request, response }
+}
